feeder/requirements.txt
pika
requests
python-firebase